Creator class created for the evolutionary alg.
This class simply defines the problem's genotype, which is defined as
BooleanGenotype. The problem is a combinatorial problem. Therefore, it
is based on selection. We have a set of variant and we want to select
the minimum number of variants that gives 100% coverage. We define a
list of variants, and each index represents a specific variant. If a
corresponding index is false, this means that the corresponding variant
is not selected (absent). Otherwise, if it is true, then it indicates
the variant is selected. This class just defines the genotype.