Objective function definition
The objective function implementation is added, which is based on the
number of selected variants and the feature coverage.

Some special conditions are added, such as if 0 variants are selected it
will set the objective function score to 0. Again if the feature
coverage is below 1 (100%), the objective function score is set to 0. We
need at least one variant, and want 100% feature coverage, but of course
the main goal is to have 100% feature coverage with minimal variants.